feat(settings): remove AI pipeline selection from settings page to simplify user experience refactor(actions): remove pipelinePreference from updateAccountSchema and related logic to streamline account updates feat(chat route): remove pipelineType from chat message processing to simplify message handling chore(migrations): add new database migrations for activity logs, chat messages, diffs, projects, subscription products, and subscriptions to support new features chore(schema): remove pipelinePreference from users schema to reflect changes in account management refactor(agent): simplify Agent class by removing pipeline handling and related logic to focus on project modifications feat(prompts): add new prompts for naive agent to enhance user interaction and context understanding chore(tests): update test for naive pipeline to reflect changes in prompt and expected output for home page modification
