+++ lib/llm/core/index.ts
@@ -1,2 +1,3 @@
-export * from './agent';
 export * from './types';
+export * from './agent';
+export * from './prompts';
