+++ lib/llm/core/prompts.ts
