--- lib/llm/pipelines/rooCode.ts
@@ -1,19 +1,0 @@
-import { Pipeline } from './types';
-
-/**
- * Implementation of a pipeline leveraging Roo Code
- */
-export class RooCodePipeline implements Pipeline {
-  async processPrompt(projectId: number, prompt: string) {
-    console.log(`🤖 Processing Roo Code pipeline for project ID: ${projectId}`);
-    console.log(`📝 Processing prompt: ${prompt.substring(0, 50)}...`);
-
-    // TODO: Implement Roo Code pipeline with actual functionality
-    // This is a placeholder for now
-
-    return {
-      success: true,
-      actions: [],
-    };
-  }
-}
