--- lib/llm/prompts/index.ts
@@ -1,2 +1,0 @@
-export * from './naive';
-export * from './rooCode';
